split mesh and spec comopnent
